feat: add category module

+++ src/banks/banks.controller.ts
@@ -3,8 +3,8 @@ import {
   Controller,
   Delete,
   Get,
+  HttpException,
   HttpStatus,
-  InternalServerErrorException,
   Param,
   Post,
   Put,
@@ -35,13 +35,8 @@ export class BankController {
     type: BankEntity,
     isArray: true,
   })
-  @ApiResponse({ status: HttpStatus.BAD_REQUEST, description: 'Bad Request' })
   async findAll() {
-    try {
-      return await this.bankService.findAll();
-    } catch (error) {
-      throw new InternalServerErrorException({ message: error });
-    }
+    return await this.bankService.findAll();
   }
 
   @Get(':id')
@@ -52,13 +47,14 @@ export class BankController {
     description: 'Success',
     type: BankEntity,
   })
-  @ApiResponse({ status: HttpStatus.BAD_REQUEST, description: 'Bad Request' })
+  @ApiResponse({ status: HttpStatus.NOT_FOUND, description: 'Bank not found' })
   async findById(@Param('id') id: string) {
-    try {
-      return await this.bankService.findById(Number(id));
-    } catch (error) {
-      throw new InternalServerErrorException({ message: error });
+    const bank = await this.bankService.findById(Number(id));
+    if (!bank) {
+      throw new HttpException('Bank not found', HttpStatus.NOT_FOUND);
     }
+
+    return bank;
   }
 
   @Post()
@@ -70,50 +66,80 @@ export class BankController {
     description: 'Success',
     type: BankEntity,
   })
-  @ApiResponse({ status: HttpStatus.BAD_REQUEST, description: 'Bad Request' })
-  async create(@Body('bank') createData: CreateBankDto) {
-    try {
-      return await this.bankService.create(createData);
-    } catch (error) {
-      throw new InternalServerErrorException({ message: error });
+  @ApiResponse({
+    status: HttpStatus.UNPROCESSABLE_ENTITY,
+    description: 'Bank with mfo exist',
+  })
+  async create(@Body() bankData: CreateBankDto) {
+    const bankByMfo = await this.bankService.findByMfo(bankData.mfo);
+    if (bankByMfo) {
+      throw new HttpException(
+        `Bank with mfo ${bankData.mfo} is exist`,
+        HttpStatus.UNPROCESSABLE_ENTITY,
+      );
     }
+
+    return await this.bankService.create(bankData);
   }
 
   @Put(':id')
   @UsePipes(new ValidationPipe())
-  @ApiOperation({ summary: 'Update a bank' })
+  @ApiOperation({ summary: 'Update a bank by specified id' })
+  @ApiParam({ name: 'id', required: true, description: 'Bank identifier' })
   @ApiBody({ type: CreateBankDto })
   @ApiResponse({
     status: HttpStatus.OK,
     description: 'Success',
     type: BankEntity,
   })
-  @ApiResponse({ status: HttpStatus.BAD_REQUEST, description: 'Bad Request' })
-  async update(
-    @Param('id') id: string,
-    @Body('bank') updateData: CreateBankDto,
-  ) {
-    try {
-      return await this.bankService.update(Number(id), updateData);
-    } catch (error) {
-      throw new InternalServerErrorException({ message: error });
+  @ApiResponse({
+    status: HttpStatus.NOT_FOUND,
+    description: 'Bank for update not found',
+  })
+  @ApiResponse({
+    status: HttpStatus.UNPROCESSABLE_ENTITY,
+    description: 'Bank with mfo exist',
+  })
+  async update(@Param('id') id: string, @Body() bankData: CreateBankDto) {
+    const bank = await this.bankService.findById(Number(id));
+    if (!bank) {
+      throw new HttpException(
+        'Bank for update not found',
+        HttpStatus.NOT_FOUND,
+      );
     }
+
+    const bankByMfo = await this.bankService.findByMfo(bankData.mfo);
+    if (bankByMfo) {
+      throw new HttpException(
+        `Bank with mfo ${bankData.mfo} is exist`,
+        HttpStatus.UNPROCESSABLE_ENTITY,
+      );
+    }
+
+    return await this.bankService.update(Number(id), bankData);
   }
 
   @Delete(':id')
-  @ApiOperation({ summary: 'Delete a bank with specified id' })
+  @ApiOperation({ summary: 'Delete a bank by specified id' })
   @ApiParam({ name: 'id', required: true, description: 'Bank identifier' })
   @ApiResponse({
     status: HttpStatus.OK,
     description: 'Success',
     type: BankEntity,
   })
-  @ApiResponse({ status: HttpStatus.BAD_REQUEST, description: 'Bad Request' })
+  @ApiResponse({
+    status: HttpStatus.NOT_FOUND,
+    description: 'Bank for delete not found',
+  })
   async delete(@Param('id') id: string) {
-    try {
-      return await this.bankService.delete(Number(id));
-    } catch (error) {
-      throw new InternalServerErrorException({ message: error });
+    const bank = await this.bankService.findById(Number(id));
+    if (!bank) {
+      throw new HttpException(
+        'Bank for delete not found',
+        HttpStatus.NOT_FOUND,
+      );
     }
+    return await this.bankService.delete(Number(id));
   }
 }
